refactor: addressed review
