+++ core/src/apps/management/recovery_device/__init__.py
@@ -10,6 +10,9 @@ if TYPE_CHECKING:
 # If set, `enforce_wordlist` must be True, because we do not support non-enforcing.
 DRY_RUN_ALLOWED_FIELDS = ("dry_run", "word_count", "enforce_wordlist", "type")
 
+def is_sdbackup_present():
+    from trezor import sdcard
+    return sdcard.is_present()
 
 async def confirm_sd_recovery(ctx: Context):
     from trezor.ui.layouts import confirm_action
@@ -43,7 +46,7 @@ async def recovery_device(ctx: Context, msg: RecoveryDevice) -> Success:
     from .homescreen import recovery_homescreen, recovery_process
     from trezor import sdcard
 
-    is_sd_present = sdcard.is_present()
+    sd_backup_present = is_sdbackup_present()
     dry_run = msg.dry_run  # local_cache_attribute
 
     # --------------------------------------------------------
@@ -69,24 +72,24 @@ async def recovery_device(ctx: Context, msg: RecoveryDevice) -> Success:
 
     # --------------------------------------------------------
     # _continue_dialog
-    if not dry_run and is_sd_present:
-        sd_restore = await confirm_sd_recovery(ctx)
-        if sd_restore:
-            restored_secret = load_sd_seed_backup()
-            storage_device.store_mnemonic_secret(
-                restored_secret,
-                BackupType.Bip39,
-                needs_backup=False,
-                no_backup=False,
-            )
-
-            await show_success(
-                ctx, "success_recovery", "You have finished recovering your wallet."
-            )
-            return Success(message="Device recovered")
-
 
     if not dry_run:
+        if sd_backup_present:
+            sd_restore = await confirm_sd_recovery(ctx)
+            if sd_restore:
+                restored_secret = load_sd_seed_backup()
+                storage_device.store_mnemonic_secret(
+                    restored_secret,
+                    BackupType.Bip39,
+                    needs_backup=False,
+                    no_backup=False,
+                )
+
+                await show_success(
+                    ctx, "success_recovery", "You have finished recovering your wallet."
+                )
+                return Success(message="Device recovered")
+
         await confirm_reset_device(ctx, "Wallet recovery", recovery=True)
     else:
         await confirm_action(
